% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{add_batch}
\alias{add_batch}
\alias{add_batch,PlateRegistry-method}
\title{Add batch to registry}
\usage{
add_batch(registry, batch, ...)

\S4method{add_batch}{PlateRegistry}(registry, batch, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{batch}{Batch object}

\item{...}{additional arguments}
}
\description{
Add batch to registry
}
\section{Methods (by class)}{
\itemize{
\item \code{add_batch(PlateRegistry)}: Add Batch object to PlateRegistry

}}
