% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{add_layout}
\alias{add_layout}
\alias{add_layout,PlateRegistry-method}
\title{Add layout to registry}
\usage{
add_layout(registry, layout_name, layout_data, ...)

\S4method{add_layout}{PlateRegistry}(registry, layout_name, layout_data, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{layout_name}{character name of layout}

\item{layout_data}{data.frame of layout}

\item{...}{additional arguments}
}
\description{
Add layout to registry
}
\section{Methods (by class)}{
\itemize{
\item \code{add_layout(PlateRegistry)}: Add layout to PlateRegistry

}}
