% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{add_plate}
\alias{add_plate}
\alias{add_plate,PlateRegistry-method}
\title{Add plate to registry}
\usage{
add_plate(registry, plate, ...)

\S4method{add_plate}{PlateRegistry}(registry, plate, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{plate}{Plate object}

\item{...}{additional arguments}
}
\description{
Add plate to registry
}
\section{Methods (by class)}{
\itemize{
\item \code{add_plate(PlateRegistry)}: Add Plate object to PlateRegistry

}}
