% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Batch-class}
\alias{Batch-class}
\title{S4 Class representing a Scoring Batch}
\description{
S4 Class representing a Scoring Batch
}
\section{Slots}{

\describe{
\item{\code{uuid}}{character.}

\item{\code{name}}{character. User-assigned name.}

\item{\code{plate_uuids}}{character. Ordered Plate UUIDs included in this batch.}

\item{\code{scores}}{list. Scoring results per strategy/metric.}

\item{\code{scoring_params}}{list. Params used in last score_batch().}

\item{\code{created_at}}{POSIXct.}
}}

