% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGExplorer-package.R
\docType{package}
\name{CGExplorer-package}
\alias{CGExplorer}
\alias{CGExplorer-package}
\title{CGExplorer: Chemo-Genomic Screening Data Preprocessor and Visualizer}
\description{
An R package providing S4 data structures, analysis methods, and a Shiny application for preprocessing, quality control, and chemo-genomic interaction scoring of 384-well plate screening datasets generated by Tecan plate readers. Plate data is managed via a persistent local registry (PlateRegistry S4 object) that accumulates across sessions and survives package upgrades.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zqzneptune/CGExplorer}
  \item Report bugs at \url{https://github.com/zqzneptune/CGExplorer/issues}
}

}
\author{
\strong{Maintainer}: Qingzhou Zhang \email{zqzneptune@hotmail.com}

Authors:
\itemize{
  \item Qingzhou Zhang \email{zqzneptune@hotmail.com}
}

}
\keyword{internal}
