% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{compute_relative_biofilm}
\alias{compute_relative_biofilm}
\title{Compute relative biofilm}
\usage{
compute_relative_biofilm(growth_m, biofilm_m)
}
\arguments{
\item{growth_m}{data.frame growth metrics}

\item{biofilm_m}{data.frame biofilm metrics}
}
\value{
data.frame joined and scored
}
\description{
Compute relative biofilm
}
