% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_biofilm_metrics}
\alias{extract_biofilm_metrics}
\title{Extract biofilm metrics}
\usage{
extract_biofilm_metrics(df)
}
\arguments{
\item{df}{data.frame corrected}
}
\value{
data.frame of biofilm metrics
}
\description{
Extract biofilm metrics
}
