% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_growth_metrics}
\alias{extract_growth_metrics}
\title{Extract growth metrics}
\usage{
extract_growth_metrics(df, growth_threshold = 0.02)
}
\arguments{
\item{df}{data.frame corrected}

\item{growth_threshold}{numeric, threshold for valid growth}
}
\value{
data.frame of metrics
}
\description{
Extract growth metrics
}
