% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{get_layout}
\alias{get_layout}
\alias{get_layout,PlateRegistry-method}
\title{Get layout from registry}
\usage{
get_layout(registry, layout_name, ...)

\S4method{get_layout}{PlateRegistry}(registry, layout_name, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{layout_name}{character name of layout}

\item{...}{additional arguments}
}
\description{
Get layout from registry
}
\section{Methods (by class)}{
\itemize{
\item \code{get_layout(PlateRegistry)}: Get layout from PlateRegistry

}}
