% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{load_registry}
\alias{load_registry}
\title{Load a PlateRegistry from RDS}
\usage{
load_registry(path)
}
\arguments{
\item{path}{character. Path to the RDS file or directory containing cgexplorer_registry.rds.}
}
\value{
A \code{\link{PlateRegistry-class}} object loaded from disk.
}
\description{
Load a PlateRegistry from RDS
}
