% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{merge_plates}
\alias{merge_plates}
\alias{merge_plates,PlateRegistry-method}
\title{Merge plates in registry}
\usage{
merge_plates(registry, uuids, ...)

\S4method{merge_plates}{PlateRegistry}(registry, uuids, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{uuids}{character vector of plate UUIDs}

\item{...}{additional arguments}
}
\description{
Merge plates in registry
}
\section{Methods (by class)}{
\itemize{
\item \code{merge_plates(PlateRegistry)}: Merge multiple plates in PlateRegistry

}}
