% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_batch}
\alias{new_batch}
\title{Constructor for Batch}
\usage{
new_batch(name, plate_uuids)
}
\arguments{
\item{name}{character. Name of the batch.}

\item{plate_uuids}{character vector. UUIDs of plates included in this batch.}
}
\value{
A \code{\link{Batch-class}} object.
}
\description{
Constructor for Batch
}
