% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{new_registry}
\alias{new_registry}
\title{Constructor for PlateRegistry}
\usage{
new_registry(project_name = "Untitled Project", project_dir = getwd())
}
\arguments{
\item{project_name}{character. Project display name.}

\item{project_dir}{character. Path to project directory storing registry.}
}
\value{
A \code{\link{PlateRegistry-class}} object.
}
\description{
Constructor for PlateRegistry
}
