% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_plate_layout}
\alias{parse_plate_layout}
\title{Parse 96 or 384 well Plate Layout}
\usage{
parse_plate_layout(filepath)
}
\arguments{
\item{filepath}{character.}
}
\value{
A list containing format info and the parsed data.frame.
}
\description{
Parse 96 or 384 well Plate Layout
}
