% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parse_tecan_txt}
\alias{parse_tecan_txt}
\title{Parse Tecan Plate Reader output (.txt)}
\usage{
parse_tecan_txt(filepath)
}
\arguments{
\item{filepath}{character.}
}
\value{
A data.frame with Plate, DateTime, Row, Column, OD
}
\description{
Parse Tecan Plate Reader output (.txt)
}
