% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PlateRegistry-class}
\alias{PlateRegistry-class}
\title{S4 Class representing the Project Plate Registry}
\description{
S4 Class representing the Project Plate Registry
}
\section{Slots}{

\describe{
\item{\code{plates}}{list. Named list: UUID -> Plate S4 object.}

\item{\code{batches}}{list. Named list: batch_name -> Batch S4 object.}

\item{\code{layouts}}{list. Named layout registry.}

\item{\code{project_name}}{character. User-assigned project name.}

\item{\code{project_dir}}{character. Absolute path to directory holding registry.rds.}

\item{\code{schema_version}}{character. Package version at save time.}

\item{\code{created_at}}{POSIXct.}

\item{\code{modified_at}}{POSIXct.}
}}

