% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_plate_qc_layout}
\alias{plot_plate_qc_layout}
\title{Plot Plate QC Layout}
\usage{
plot_plate_qc_layout(
  plate,
  metric = "Growth_Endpoint_OD",
  colorscale = "plasma",
  cmin = NULL,
  cmax = NULL
)
}
\arguments{
\item{plate}{Plate object}

\item{metric}{character. Column name in metrics to plot.}

\item{colorscale}{character. Color scheme key.}

\item{cmin}{numeric. Minimum value for color scale.}

\item{cmax}{numeric. Maximum value for color scale.}
}
\value{
plotly object
}
\description{
Plot Plate QC Layout
}
