% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_replicate_growth_curves}
\alias{plot_replicate_growth_curves}
\title{Plot Replicate Growth Curves}
\usage{
plot_replicate_growth_curves(replicate_data, stain_hr = 26)
}
\arguments{
\item{replicate_data}{data.frame containing Time_Hours, OD_Raw or OD_Corrected, Row, Column, Gene, Rep_Suffix}

\item{stain_hr}{numeric staining hour threshold}
}
\value{
plotly object
}
\description{
Plot Replicate Growth Curves
}
