% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{remove_batch}
\alias{remove_batch}
\alias{remove_batch,PlateRegistry-method}
\title{Remove batch from registry}
\usage{
remove_batch(registry, batch_name, ...)

\S4method{remove_batch}{PlateRegistry}(registry, batch_name, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{batch_name}{character batch name}

\item{...}{additional arguments}
}
\description{
Remove batch from registry
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_batch(PlateRegistry)}: Remove Batch object from PlateRegistry

}}
