% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{remove_layout}
\alias{remove_layout}
\alias{remove_layout,PlateRegistry-method}
\title{Remove layout from registry}
\usage{
remove_layout(registry, layout_name, ...)

\S4method{remove_layout}{PlateRegistry}(registry, layout_name, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{layout_name}{character name of layout}

\item{...}{additional arguments}
}
\description{
Remove layout from registry
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_layout(PlateRegistry)}: Remove layout from PlateRegistry

}}
