% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{remove_plate}
\alias{remove_plate}
\alias{remove_plate,PlateRegistry-method}
\title{Remove plate from registry}
\usage{
remove_plate(registry, uuid, ...)

\S4method{remove_plate}{PlateRegistry}(registry, uuid, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{uuid}{character plate UUID}

\item{...}{additional arguments}
}
\description{
Remove plate from registry
}
\section{Methods (by class)}{
\itemize{
\item \code{remove_plate(PlateRegistry)}: Remove Plate object from PlateRegistry

}}
