% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_cgexplorer_app}
\alias{run_cgexplorer_app}
\title{Run the Shiny Application}
\usage{
run_cgexplorer_app(...)
}
\arguments{
\item{...}{arguments to pass to golem_opts.}
}
\description{
Run the Shiny Application
}
