% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{save_registry}
\alias{save_registry}
\alias{save_registry,PlateRegistry-method}
\title{Save registry to disk}
\usage{
save_registry(registry, ...)

\S4method{save_registry}{PlateRegistry}(registry, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{...}{additional arguments}
}
\description{
Save registry to disk
}
\section{Methods (by class)}{
\itemize{
\item \code{save_registry(PlateRegistry)}: Save PlateRegistry to RDS file

}}
