% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{score_batch_dataset}
\alias{score_batch_dataset}
\title{Standardized Scoring Engine}
\usage{
score_batch_dataset(
  pheno_df,
  strategy_type = c("NoDrug_Plates", "NoDrug_Wells"),
  metrics,
  control_drug = "NoDrug"
)
}
\arguments{
\item{pheno_df}{data.frame containing aggregated phenotypic metrics across a batch}

\item{strategy_type}{character, either "NoDrug_Plates" or "NoDrug_Wells"}

\item{metrics}{character vector of target metrics}

\item{control_drug}{character, name of control treatment}
}
\value{
data.frame of scores
}
\description{
Standardized Scoring Engine
}
