% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{subtract_plate_blanks}
\alias{subtract_plate_blanks}
\title{Subtract blanks at the plate level}
\usage{
subtract_plate_blanks(df, blank_threshold = 0.05)
}
\arguments{
\item{df}{data.frame with Row, Column, Time_Hours, OD_Raw, Type, Is_Stained}

\item{blank_threshold}{numeric, threshold for bad blanks}
}
\value{
list with corrected `data` and `removed_blanks`
}
\description{
Subtract blanks at the plate level
}
