% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods.R
\name{update_plate}
\alias{update_plate}
\alias{update_plate,PlateRegistry-method}
\title{Update plate in registry}
\usage{
update_plate(registry, plate, ...)

\S4method{update_plate}{PlateRegistry}(registry, plate, ...)
}
\arguments{
\item{registry}{PlateRegistry object}

\item{plate}{Plate object}

\item{...}{additional arguments}
}
\description{
Update plate in registry
}
\section{Methods (by class)}{
\itemize{
\item \code{update_plate(PlateRegistry)}: Update Plate object in PlateRegistry

}}
